+++ tests/bluetooth/shell/src/main.c
@@ -758,7 +758,7 @@ done:
 
 static struct bt_gatt_read_params read_params;
 
-static uint8_t read_func(struct bt_conn *conn, int err,
+static uint8_t read_func(struct bt_conn *conn, uint8_t err,
 			 struct bt_gatt_read_params *params,
 			 const void *data, uint16_t length)
 {
